(PC-17132)[PRO] refactor: remove venue unused component
