+++ pro/src/components/pages/Home/Venues/__specs__/Venue.spec.jsx
@@ -1,12 +1,15 @@
 import '@testing-library/jest-dom'
 
-import { act, render, screen, within } from '@testing-library/react'
+import { act, render, screen, waitFor, within } from '@testing-library/react'
+import userEvent from '@testing-library/user-event'
 import React from 'react'
 import { Provider } from 'react-redux'
 import { MemoryRouter } from 'react-router'
 
 import Venue from 'components/pages/Home/Venues/Venue'
+import * as pcapi from 'repository/pcapi/pcapi'
 import { configureTestStore } from 'store/testUtils'
+import { loadFakeApiVenueStats } from 'utils/fakeApi'
 
 const renderVenue = async (props, store) => {
   return await act(async () => {
@@ -31,27 +34,35 @@ describe('venues', () => {
       isVirtual: false,
       name: 'My venue',
       offererId: 'OFFERER01',
-      venueStats: {
-        activeBookingsQuantity: 0,
-        activeOffersCount: 2,
-        soldOutOffersCount: 3,
-        validatedBookingsQuantity: 1,
-      },
     }
+    loadFakeApiVenueStats({
+      activeBookingsQuantity: 0,
+      activeOffersCount: 2,
+      soldOutOffersCount: 3,
+      validatedBookingsQuantity: 1,
+    })
   })
 
   it('should display stats tiles', async () => {
     // When
     await renderVenue(props, store)
 
+    await userEvent.click(screen.getByRole('button', { name: 'Afficher' }))
+
     // Then
+    expect(pcapi.getVenueStats).toHaveBeenCalledWith(props.id)
+
     const [
       activeOffersStat,
       activeBookingsStat,
       validatedBookingsStat,
       outOfStockOffersStat,
     ] = screen.getAllByTestId('venue-stat')
-    expect(within(activeOffersStat).getByText('2')).toBeInTheDocument()
+
+    await waitFor(() =>
+      expect(within(activeOffersStat).getByText('2')).toBeInTheDocument()
+    )
+
     expect(
       within(activeOffersStat).getByText('Offres publiées')
     ).toBeInTheDocument()
@@ -75,6 +86,11 @@ describe('venues', () => {
   it('should contain a link for each stats', async () => {
     // When
     await renderVenue(props, store)
+    await userEvent.click(screen.getByRole('button', { name: 'Afficher' }))
+    const [activeOffersStat] = screen.getAllByTestId('venue-stat')
+    await waitFor(() =>
+      expect(within(activeOffersStat).getByText('2')).toBeInTheDocument()
+    )
 
     // Then
     expect(screen.getAllByRole('link', { name: 'Voir' })).toHaveLength(4)
@@ -83,7 +99,7 @@ describe('venues', () => {
   it('should redirect to filtered bookings when clicking on link', async () => {
     // When
     await renderVenue(props, store)
-
+    await userEvent.click(screen.getByRole('button', { name: 'Afficher' }))
     // Then
     const [
       activeOffersStat,
@@ -91,16 +107,24 @@ describe('venues', () => {
       validatedBookingsStat,
       outOfStockOffersStat,
     ] = screen.getAllByTestId('venue-stat')
+
+    await waitFor(() =>
+      expect(within(activeOffersStat).getByText('2')).toBeInTheDocument()
+    )
+
     expect(
       within(activeOffersStat).getByRole('link', { name: 'Voir' })
     ).toHaveAttribute('href', '/offres?lieu=VENUE01&statut=active')
     const byRole = within(validatedBookingsStat).getByRole('link', {
       name: 'Voir',
     })
-    expect(byRole).toHaveAttribute('href', '/reservations')
+    expect(byRole).toHaveAttribute(
+      'href',
+      '/reservations?page=1&bookingStatusFilter=validated&offerVenueId=VENUE01'
+    )
     expect(
       within(activeBookingsStat).getByRole('link', { name: 'Voir' })
-    ).toHaveAttribute('href', '/reservations')
+    ).toHaveAttribute('href', '/reservations?page=1&offerVenueId=VENUE01')
     expect(
       within(outOfStockOffersStat).getByRole('link', { name: 'Voir' })
     ).toHaveAttribute('href', '/offres?lieu=VENUE01&statut=epuisee')
@@ -113,16 +137,58 @@ describe('venues', () => {
 
       // When
       await renderVenue(props, store)
+      await userEvent.click(screen.getByTitle('Afficher'))
+      const [activeOffersStat] = screen.getAllByTestId('venue-stat')
+      await waitFor(() =>
+        expect(within(activeOffersStat).getByText('2')).toBeInTheDocument()
+      )
 
       // Then
       expect(
         screen.getByRole('link', { name: 'Créer une nouvelle offre numérique' })
       ).toBeInTheDocument()
     })
+  })
+
+  describe('physical venue section', () => {
+    it('should display create offer link', async () => {
+      // Given
+      props.isVirtual = false
+
+      // When
+      await renderVenue(props, store)
+      await userEvent.click(screen.getByTitle('Afficher'))
+      const [activeOffersStat] = screen.getAllByTestId('venue-stat')
+      await waitFor(() =>
+        expect(within(activeOffersStat).getByText('2')).toBeInTheDocument()
+      )
+
+      // Then
+      expect(
+        screen.getByRole('link', { name: 'Créer une nouvelle offre' })
+      ).toBeInTheDocument()
+    })
+
+    it('should display edition venue link', async () => {
+      // Given
+      props.isVirtual = false
+
+      // When
+      await renderVenue(props, store)
+      await userEvent.click(screen.getByRole('button', { name: 'Afficher' }))
+      const [activeOffersStat] = screen.getAllByTestId('venue-stat')
+      await waitFor(() =>
+        expect(within(activeOffersStat).getByText('2')).toBeInTheDocument()
+      )
+
+      // Then
+      expect(screen.getByRole('link', { name: 'Modifier' }).href).toBe(
+        'http://localhost/structures/OFFERER01/lieux/VENUE01?modification'
+      )
+    })
 
     it('should display add bank information when venue does not have a business unit', async () => {
       // Given
-      props.isVirtual = true
       props.hasBusinessUnit = false
       const storeOverrides = configureTestStore({
         features: {
@@ -140,42 +206,17 @@ describe('venues', () => {
         'http://localhost/structures/OFFERER01/lieux/VENUE01?modification'
       )
     })
-  })
 
-  describe('physical venue section', () => {
-    it('should display create offer link', async () => {
+    it('should display add bank information when venue does not have a reimbursement point', async () => {
       // Given
-      props.isVirtual = false
-
-      // When
-      await renderVenue(props, store)
-
-      // Then
-      expect(
-        screen.getByRole('link', { name: 'Créer une nouvelle offre' })
-      ).toBeInTheDocument()
-    })
-
-    it('should display edition venue link', async () => {
-      // Given
-      props.isVirtual = false
-
-      // When
-      await renderVenue(props, store)
-
-      // Then
-      expect(screen.getByRole('link', { name: 'Modifier' }).href).toBe(
-        'http://localhost/structures/OFFERER01/lieux/VENUE01?modification'
-      )
-    })
-
-    it('should display add bank information when venue does not have a business unit', async () => {
-      // Given
-      props.hasBusinessUnit = false
+      props.hasMissingReimbursementPoint = true
       const storeOverrides = configureTestStore({
         features: {
           list: [
-            { isActive: true, nameKey: 'ENFORCE_BANK_INFORMATION_WITH_SIRET' },
+            {
+              isActive: true,
+              nameKey: 'ENABLE_NEW_BANK_INFORMATIONS_CREATION',
+            },
           ],
         },
       })
